// src/app/page/master/unit-bagian/unit-bagian.interface.ts
import { EventEmitter } from '@angular/core';

export interface  UnitBagian{
    kode?: any;
    kdGolonganPegawai?: number;
    hargaSatuan?: string
    kdOperatorRate?: string;
    kdKomponenHarga?: string;
    status1?: string;
    noRec?: string;
    label?: any;
    value?: any;
    namaStatus?: string;
    id?: any;
    awardOffence?: any;
}